+++ UpstraUIKit/UpstraUIKit/Modules/Comunity/Community Notifications/CompressedSizeController.swift
@@ -1,0 +1,36 @@
+//
+// Created by Leke Abolade
+// Copyright ® 2023 Noom, Inc. All Rights Reserved
+//
+
+import UIKit
+
+final class CompressedSizeController: UIViewController {
+
+    private let contentViewController: UIViewController
+
+    init(contentViewController: UIViewController) {
+        self.contentViewController = contentViewController
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.view.translatesAutoresizingMaskIntoConstraints = false
+        //self.noom.addChildFittingSuperview(contentViewController)
+
+        self.preferredContentSize = self.view.systemLayoutSizeFitting(UIView.layoutFittingCompressedSize)
+        self.view.setNeedsLayout()
+        self.view.layoutIfNeeded()
+    }
+
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        preferredContentSize = view.systemLayoutSizeFitting(UIView.layoutFittingCompressedSize)
+    }
+}
